Fix: status codes
- handle illegal argument exception - return 400 bad request
- handle 4xx status code
- minor changes to by in sync with documentation

+++ src/main/java/com/babakjan/moneybag/service/CategoryService.java
@@ -77,6 +77,9 @@ public class CategoryService {
         if (id == null) {
             throw new CategoryNotFoundException("Category id can't be null.");
         }
+        if (categoryRepository.findById(id).isEmpty()) {
+            throw new CategoryNotFoundException(id);
+        }
         categoryRepository.deleteById(id);
     }
 
@@ -110,7 +113,7 @@ public class CategoryService {
     }
 
     /**
-     * Get analytic of category.
+     * Get analytic of categories. Role ADMIN can access analytic of all users, role USER only of their accounts.
      * @param userId user id (analytic of records of this user)
      * @param dateGe date greater or equal than (inclusive)
      * @param dateLt date lower than (exclusive)
